keep data type of iterator range as double, could overflow int

+++ Vcw.Core/source/VirtualCamera.cpp
@@ -185,6 +185,8 @@ namespace Vcw
 
 			cameraPerspectiveLimits[k] = cameraPointTop_px;
 			cameraPerspectiveLimits[k + PropImageSize.width] = cameraPointBottom_px;
+
+			//std::cout << cameraPointTop << ", " <<  cameraPointBottom << "\n";
 		}
 
 		for (int k = 1; k < PropImageSize.height - 1; k++)
@@ -200,10 +202,12 @@ namespace Vcw
 
 			cameraPerspectiveLimits[k + 2 * PropImageSize.width - 1] = cameraPointLeft_px;
 			cameraPerspectiveLimits[k + 2 * PropImageSize.width + PropImageSize.height - 3] = cameraPointRight_px;
+
+			//std::cout << cameraPointLeft << ", " << cameraPointRight << "\n";
 		}
 
 		/*
-		for(int k=0; k< cameraPerspectiveLimits.size();k++)
+		for (int k = 0; k < cameraPerspectiveLimits.size(); k++)
 		{
 			std::cout << cameraPerspectiveLimits[k] << "\n";
 		}
@@ -213,8 +217,12 @@ namespace Vcw
 		std::tie(minX, maxX) = std::minmax_element(begin(cameraPerspectiveLimits), std::end(cameraPerspectiveLimits), [](cv::Point2d const& p0, cv::Point2d const& p1) {return p0.x < p1.x; });
 		std::tie(minY, maxY) = std::minmax_element(begin(cameraPerspectiveLimits), std::end(cameraPerspectiveLimits), [](cv::Point2d const& p0, cv::Point2d const& p1) {return p0.y < p1.y; });
 
-		const cv::Range RangeX(cv::max(0, (int)floor(minX->x)), cv::min(cameraProperties.Resolution.width - 1, (int)ceil(maxX->x)));
-		const cv::Range RangeY(cv::max(0, (int)floor(minY->y)), cv::min(cameraProperties.Resolution.height - 1, (int)ceil(maxY->y)));
+		const cv::Range RangeX(cv::max(0.0, floor(minX->x)), cv::min(cameraProperties.Resolution.width - 1.0, ceil(maxX->x)));
+		const cv::Range RangeY(cv::max(0.0, floor(minY->y)), cv::min(cameraProperties.Resolution.height - 1.0, ceil(maxY->y)));
+
+		//std::cout << minX->x << ", " << maxX->x << ", " << minY->y << ", " << maxY->y << "\n";
+
+		//std::cout << RangeX.start << ", " << RangeX.end << ", " << RangeY.start << ", " << RangeY.end << "\n";
 
 		if (RangeX.size() < 0) return;
 		if (RangeY.size() < 0) return;
